backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.52.0
pydantic==2.9.2
python-dotenv==1.0.1
qdrant-client==1.12.0